hoa6

+++ PBL3/DTO/NguyenLieu.cs
@@ -23,7 +23,7 @@ namespace PBL3.DTO
     
         public int MaNL { get; set; }
         public string TenNL { get; set; }
-        public Nullable<int> SLTonKho { get; set; }
+        public Nullable<decimal> SLTonKho { get; set; }
         public string DonViTinh { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
